GUACAMOLE-1289: Move AuthenticationSession components to guacamole-exit.
